div contenteditable added. Also speedups uncommented
Spent a ton of time thinking about performance.
There's an odd thing that focus() takes a bunch of time,
but when divs are contentEditable it goes a bit faster.
It's also true that if I use faster immutable data structures,
the shouldcomponentupdate will run faster. That right now is the largest
bottleneck (although it's 15%.. so it's not overwhelming).

It looks like if I want to support >1500 on the page I'll need to implement
constant time operations, like moving the cursor without "redrawing" with
react. Those in memory computations take too long :/