Fixed html entities bugs
Took me foreeeever to do this, because I was thinking how to do bold, etc.
and I was kinda overcomplicating it, not finding the elegant solution.

Specifically, I thought I was going to have problems with where the cursor would
go, because there would be some htmlentities in the tree datastructure that are not shown,
so the lengths would be different. Buttt, the nice hacky solution is to just hold on
to the html in my tree structure, and _always_ turn it into htmlentities when drawing it.
This means no html for now, but it's an easy solution!