Speed up rendering by tracking diffs of changes
Before this change, React was going through every single element and
doing a DOM difference check, because the props were changing. The props
were changing because the props have a huge Node, and comparing those
seems to always be false.

Now, I'm speeding things up using shouldComponentUpdate. For every
operation to the tree, I'm tracking what diff happens. I then add that
to a diff list, and then shouldComponentUpdate checks that diff list.
This results in huge speed improvements.

There are still some places where I'm just deciding to rerender
everything. This is during:
- Zooming
- Initial page load
- Indending
- Expand/contracting
- Completing (Note, this has to go through all the children and put a
strike through on each)
- Deletions when everything is empty (this isn't slow, because the doc
is empty)

I've noticed that when doing these full loads, the operations are
substantially slower than Workflowy. The reason for this is that
Workflowy "cheats" by not rendering the full DOM to the page. Just the
parts that you expect to see. So they don't have to handle so much DOM
at any given time. It's a cool hack that they do, but it has
limitations. For example, "cmd+f" on workflowy won't properly show you
results unless you scroll down all the way to the end of the document.

Mini bug fixes